pyproject.toml
[tool.poetry]
description = "Imagen de Django 2.2.13 con Nginx y Apache Airflow."
name = "carnetdevacunacion"
version = "0.1.0"

license = "Apache-2.0"

authors = [
  "Saúl Galán <saul.galan@continuum.com>",
  "Nicolas Caballero <nicolas.caballero@continuum.com>",
]

readme = 'README.md' # Markdown files are supported

homepage = "https://github.com/continuum/carnetdevacunacion"
repository = "https://github.com/continuum/carnetdevacunacion"

keywords = ['packaging', 'poetry']

[tool.poetry.dependencies]
# https://python-poetry.org/docs/dependency-specification/
Django = "^3.0"
asgiref = "^3.2"
django-cors-headers = "^3.5"
django-extensions = "^3.0"
django-redis = "^4.12"
django-rest-auth = "^0.9"
djangorestframework = "^3.10"
djangotoolbox = "^1.8"
nose = "^1.3"
psycopg2 = "^2.7.4"
python = "~3.7"
python-dateutil = "^2.8"
python-dotenv = "^0.15"
python-slugify = "^4.0"
redis = "^3.2"
requests = "^2.22"
termcolor = "^1.1"
urllib3 = "^1.24.1"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5"
flake8 = "^3.8"
flake8-commas = "^2.0"
ipykernel = "^5.3"
ipython = "^7.17"
ipywidgets = "^7.5"
isort = "^5.6"
jedi = "^0.17"
jupyter = "^1.0"
jupyter-client = "^6.1"
jupyter-console = "^6.1"
jupyter-core = "^4.6"
notebook = "^6.1"
qtconsole = "^4.7"
widgetsnbextension = "^3.5"
